Crate ralloc

Ralloc: The memory efficient allocator.

This crates define the user space allocator for Redox, which emphasizes performance and memory efficiency.

Ralloc seems to reimplement everything. Why?

Memory allocators cannot depend on libraries or primitives, which allocates. This is a relatively strong condition, which means that you are forced to rewrite primitives and make sure no allocation ever happens.

Macros

bk_log
Log with bookkeeper data to the appropriate source.

log
Log to the appropriate source.

Functions

alloc
Allocate a block of memory.

free⚠
Free a buffer.

realloc⚠
Reallocate memory.

realloc_inplace⚠
Try to reallocate the buffer inplace.

sbrk⚠
SBRK symbol which can coexist with the allocator.

set_oom_handler
Set the OOM handler.